Phuongthe12

+++ src/main/java/luckystore/datn/model/response/KichThuocResponse.java
@@ -16,9 +16,9 @@ public class KichThuocResponse {
 
     private String ten;
 
-    private Float chieuDai;
+    private Double chieuDai;
 
-    private Float chieuRong;
+    private Double chieuRong;
 
     private String moTa;
